IPD-40 #comment added iconik authorizer and example of use authorizer
