--- authorizers/iconik/context.ts
@@ -1,14 +1,0 @@
-export interface IconikContext {
-  appId: string;
-  systemDomainId: string;
-  caller: {
-    id: string;
-    email: string;
-  };
-  appOwner: {
-    id: string;
-    email: string;
-  };
-  authToken: string;
-  iconikUrl: string;
-}
