+++ authorizers/iconik/interfaces/context.ts
@@ -1,0 +1,14 @@
+export interface IconikContext {
+  appId: string;
+  systemDomainId: string;
+  caller: {
+    id: string;
+    email: string;
+  };
+  appOwner: {
+    id: string;
+    email: string;
+  };
+  authToken: string;
+  iconikUrl: string;
+}
